Fix the test suite and add docker setup
There are quite some code violations that was failing the tests,
this commit tries to fix those issues, and then it also updates
the rake task to be more gentle for now.

This commit also adds `docker` setup for development, this is a
personal preference, but I believe it would help any developer
to easily run and do development on this project.